4/27 commit

+++ src/main/kotlin/bj9461.kt
@@ -1,11 +1,11 @@
 fun main() {
-    val num = readln().toInt()
-    for (j in 0 until num) {
-        var arr = arrayListOf(1, 1, 1, 2, 2, 3, 4, 5, 7, 9)
-        val n = readln().toInt()
+    val num = readln().toLong()
+    for (j in 0L until num) {
+        var arr = arrayListOf(1L, 1L, 1L, 2L, 2L, 3L, 4L, 5L, 7L, 9L)
+        val n = readln().toLong()
         for (i in arr.size - 1..n) {
-            arr.add(arr[i] + arr[i - 4])
+            arr.add(arr[i.toInt()] + arr[i.toInt() - 4])
         }
-        println(arr[n - 1])
+        println(arr[n.toInt() - 1])
     }
 }
